+++ sdk/src/main/java/com/iovation/launchkey/sdk/transport/domain/AuthPolicy.java
@@ -16,10 +16,7 @@ import com.fasterxml.jackson.annotation.*;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
@@ -54,33 +51,62 @@ public class AuthPolicy {
         this.deviceIntegrity = deviceIntegrity;
     }
 
-    @SuppressWarnings("unchecked")
     @JsonCreator
     public AuthPolicy(@JsonProperty("minimum_requirements") List<MinimumRequirement> minimumRequirements,
-                      @JsonProperty("factors") ArrayNode factors) {
-        this.minimumRequirements = minimumRequirements;
-        geoFenceLocations = new ArrayList<>();
-        Boolean deviceIntegrity = null;
-        try {
-            for (JsonNode factor : factors) {
-                final String factorName = factor.get("factor").asText();
-                final JsonNode attributes = factor.get("attributes");
-                if (factorName.equals("geofence")) {
-                    for (JsonNode location : attributes.get("locations")) {
-                        String name = location.has("name") ? location.get("name").asText() : null;
-                        addGeoFence(
-                                name,
-                                location.get("radius").asDouble(),
-                                location.get("latitude").asDouble(),
-                                location.get("longitude").asDouble());
-                    }
-                } else if (factorName.equals("device integrity")) {
-                    deviceIntegrity = attributes.get("factor enabled").asInt() == 1;
-                }
+                      @JsonProperty("factors") ArrayNode factors, @JsonProperty("types") List<String> types,
+                      @JsonProperty("amount") Integer amount,
+                      @JsonProperty("geofences") List<AuthPolicy.Location> geofences) {
+        if (minimumRequirements == null) {
+            geoFenceLocations = geofences;
+            deviceIntegrity = null;
+            Integer knowledge;
+            Integer possession;
+            Integer inherence;
+            if (types == null) {
+                knowledge = null;
+                possession = null;
+                inherence = null;
+            } else {
+                knowledge = types.contains("knowledge") ? ONE : ZERO;
+                possession = types.contains("possession") ? ONE : ZERO;
+                inherence = types.contains("inherence") ? ONE : ZERO;
             }
-            this.deviceIntegrity = deviceIntegrity;
-        } catch (ClassCastException | NullPointerException e) {
-            throw new IllegalArgumentException("The argument \"factors\" does not conform to the specification", e);
+            this.minimumRequirements = new ArrayList<>();
+            MinimumRequirement minimumRequirement = new MinimumRequirement(
+                    MinimumRequirement.Type.AUTHENTICATED,
+                    amount,
+                    knowledge,
+                    inherence,
+                    possession
+            );
+            this.minimumRequirements.add(minimumRequirement);
+
+
+        } else{
+            this.minimumRequirements = minimumRequirements;
+            geoFenceLocations = new ArrayList<>();
+            Boolean deviceIntegrity = null;
+            try {
+                for (JsonNode factor : factors) {
+                    final String factorName = factor.get("factor").asText();
+                    final JsonNode attributes = factor.get("attributes");
+                    if (factorName.equals("geofence")) {
+                        for (JsonNode location : attributes.get("locations")) {
+                            String name = location.has("name") ? location.get("name").asText() : null;
+                            addGeoFence(
+                                    name,
+                                    location.get("radius").asDouble(),
+                                    location.get("latitude").asDouble(),
+                                    location.get("longitude").asDouble());
+                        }
+                    } else if (factorName.equals("device integrity")) {
+                        deviceIntegrity = attributes.get("factor enabled").asInt() == 1;
+                    }
+                }
+                this.deviceIntegrity = deviceIntegrity;
+            } catch (ClassCastException | NullPointerException e) {
+                throw new IllegalArgumentException("The argument \"factors\" does not conform to the specification", e);
+            }
         }
     }
 
@@ -127,37 +153,32 @@ public class AuthPolicy {
         return factors;
     }
 
+    public Boolean getDeviceIntegrity() {
+        return deviceIntegrity;
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (!(o instanceof AuthPolicy)) return false;
-
         AuthPolicy that = (AuthPolicy) o;
-
-        if (getMinimumRequirements() != null ? !getMinimumRequirements().equals(that.getMinimumRequirements()) :
-                that.getMinimumRequirements() != null)
-            return false;
-        return geoFenceLocations != null ? geoFenceLocations.equals(that.geoFenceLocations) :
-                that.geoFenceLocations == null;
+        return Objects.equals(getMinimumRequirements(), that.getMinimumRequirements()) &&
+                Objects.equals(geoFenceLocations, that.geoFenceLocations) &&
+                Objects.equals(getDeviceIntegrity(), that.getDeviceIntegrity());
     }
 
     @Override
     public int hashCode() {
-        int result = getMinimumRequirements() != null ? getMinimumRequirements().hashCode() : 0;
-        result = 31 * result + (geoFenceLocations != null ? geoFenceLocations.hashCode() : 0);
-        return result;
+        return Objects.hash(getMinimumRequirements(), geoFenceLocations, getDeviceIntegrity());
     }
 
     @Override
     public String toString() {
-        return this.getClass().getSimpleName() + "{" +
+        return "AuthPolicy{" +
                 "minimumRequirements=" + minimumRequirements +
                 ", geoFenceLocations=" + geoFenceLocations +
+                ", deviceIntegrity=" + deviceIntegrity +
                 '}';
-    }
-
-    public Boolean getDeviceIntegrity() {
-        return deviceIntegrity;
     }
 
     @JsonIgnoreProperties(ignoreUnknown = true)
@@ -267,7 +288,9 @@ public class AuthPolicy {
         private final double latitude;
         private final double longitude;
 
-        public Location(String name, double radius, double latitude, double longitude) {
+        @JsonCreator
+        public Location(@JsonProperty("name") String name, @JsonProperty("radius") double radius,
+                        @JsonProperty("latitude") double latitude, @JsonProperty("longitude") double longitude) {
             this.name = name;
             this.radius = radius;
             this.latitude = latitude;